start restructure task2 and new migrations
