+++ task2/alembic.ini
@@ -60,8 +60,10 @@ version_path_separator = os  # Use os.pathsep. Default configuration used for ne
 # are written from script.py.mako
 # output_encoding = utf-8
 
-# sqlalchemy.url = postgresql://postgres:postgres@localhost:5432/postgres1  local pc
-sqlalchemy.url = postgresql://postgres:postgres@db:5432/postgres
+# local pc 
+sqlalchemy.url = postgresql://postgres:postgres@localhost:5432/postgres 
+# container db
+# sqlalchemy.url = postgresql://postgres:postgres@db:5432/postgres
 
 
 [post_write_hooks]
