+++ task2/src/audio/models.py
@@ -1,27 +1,16 @@
 import uuid
 
-from sqlalchemy import UUID, Column, ForeignKey, String
-from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
-from sqlalchemy.orm import sessionmaker
+from sqlalchemy import UUID, Column, ForeignKey, MetaData, String
 
-from auth.models import Base
+from auth.models import User
+from src.database import Base
 
-# import settings
-
-
-# # create async engine for interaction with database
-# engine = create_async_engine(
-#     settings.REAL_DATABASE_URL, future=True, echo=True
-# )
-
-# # create session for the interaction with database
-# async_session = sessionmaker(
-#     engine, expire_on_commit=False, class_=AsyncSession
-# )
+metadata = MetaData()
 
 
 class Audio(Base):
     __tablename__ = "audiofile"
+    metadata = metadata
     id: UUID = Column(
         UUID(as_uuid=True),
         nullable=False,
@@ -30,4 +19,4 @@ class Audio(Base):
     )
     filename: str = Column(String, nullable=False, default=None)
     filepath: str = Column(String, nullable=False, default=None)
-    user_id: str = Column(UUID(as_uuid=True), ForeignKey("user.id"))
+    user_id: str = Column(UUID(as_uuid=True), ForeignKey(User.id))
