+++ task2/src/audio/resoponses.py
@@ -1,0 +1,12 @@
+responses = {
+    400: {
+        "description": "Не верный формат записи",
+        "content": {
+            "application/json": {
+                "example": {
+                    "detail": "Убедитесь, что загружаете файл в формате .wav"
+                }
+            }
+        },
+    }
+}
