+++ task2/src/auth/base_config.py
@@ -11,7 +11,7 @@ SECRET = "SECRET"
 
 
 def get_jwt_strategy() -> JWTStrategy:
-    return JWTStrategy(secret=SECRET, lifetime_seconds=3600 * 24)
+    return JWTStrategy(secret=SECRET, lifetime_seconds=(3600 * 24))
 
 
 auth_backend = AuthenticationBackend(
