+++ task2/src/auth/models.py
@@ -6,59 +6,30 @@ from fastapi_users.db import (
     SQLAlchemyBaseUserTableUUID,
     SQLAlchemyUserDatabase,
 )
-from sqlalchemy import Boolean, String
-from sqlalchemy.ext.asyncio import (
-    AsyncSession,
-    async_sessionmaker,
-    create_async_engine,
-)
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
+from sqlalchemy import MetaData, String
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.orm import Mapped, mapped_column
 
-from database import async_session_maker
-from settings import REAL_DATABASE_URL
+from src.database import Base, async_session_maker
+from src.settings import REAL_DATABASE_URL
 
 DATABASE_URL = REAL_DATABASE_URL
 UUID_ID = uuid.UUID
 
 
-class Base(DeclarativeBase):
-    pass
+metadata = MetaData()
 
 
 class User(SQLAlchemyBaseUserTableUUID, Base):
     __tablename__ = "user"
+    metadata = metadata
 
     if TYPE_CHECKING:  #   pragma: no cover
-        id: UUID_ID
-        email: str
-        hashed_password: str
-        is_active: bool
-        is_superuser: bool
-        is_verified: bool
         username: str
     else:
-        email: Mapped[str] = mapped_column(
-            String(length=320), unique=True, index=True, nullable=False
-        )
-        hashed_password: Mapped[str] = mapped_column(
-            String(length=1024), nullable=False
-        )
-        is_active: Mapped[bool] = mapped_column(
-            Boolean, default=True, nullable=False
-        )
-        is_superuser: Mapped[bool] = mapped_column(
-            Boolean, default=False, nullable=False
-        )
-        is_verified: Mapped[bool] = mapped_column(
-            Boolean, default=False, nullable=False
-        )
         username: Mapped[bool] = mapped_column(
             String(length=50), unique=True, default=False, nullable=False
         )
-
-
-# engine = create_async_engine(DATABASE_URL)
-# async_session_maker = async_sessionmaker(engine, expire_on_commit=False)
 
 
 async def get_async_session() -> AsyncGenerator[AsyncSession, None]:
